Fix circle dependency
